+++ include/graphic/components.hpp
@@ -14,11 +14,11 @@
 namespace Graphic
 {
 
-Object& addSphere(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation, float radius, Color color);
-Object& addPlane(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation, float height, float width, Color color);
-Object& addRectangular(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation, const Eigen::Vector3f& size, Color color);
-Object& addCylinder(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation, float radius, float height, Color color);
-Object& addTeapot(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation, int size, Color color);
-Object& addSTLModel(const Eigen::Vector3f& position, const Eigen::Vector3f& rotation, const std::string filepath, bool colored, Color color);
+Object& addSphere(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation, double radius, Color color);
+Object& addPlane(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation, double height, double width, Color color);
+Object& addRectangular(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation, const Eigen::Vector3d& size, Color color);
+Object& addCylinder(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation, double radius, double height, Color color);
+Object& addTeapot(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation, int size, Color color);
+Object& addSTLModel(const Eigen::Vector3d& position, const Eigen::Vector3d& rotation, const std::string filepath, bool colored, Color color);
 
 }  // namespace Graphic
