+++ include/graphic/impl/stl/binary.hpp
@@ -16,7 +16,7 @@ struct Binary {
 
     std::vector<Polygon> polygons;
 
-    float parse_float(std::ifstream& file)
+    double parse_float(std::ifstream& file)
     {
         union {
             char buf[4];
@@ -26,9 +26,9 @@ struct Binary {
         return data.val;
     }
 
-    Eigen::Vector3f parse_point(std::ifstream& file)
+    Eigen::Vector3d parse_point(std::ifstream& file)
     {
-        return Eigen::Vector3f{
+        return Eigen::Vector3d{
             parse_float(file),
             parse_float(file),
             parse_float(file)};
